Moving interfaces into different namespace
